# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinCommand Executive Treasury API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List all vendors/beneficiaries
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vendors
          content:
            application/json:
              schema:
                type: object
                properties:
                  vendors:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vendor"
                required: [vendors]
    post:
      operationId: createVendor
      tags: [vendors]
      summary: Create a new vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorInput"
      responses:
        "201":
          description: Created vendor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      summary: Get vendor profile with running balance and history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vendor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorDetail"
  /vendors/{id}/liabilities:
    get:
      operationId: getVendorLiabilities
      tags: [vendors]
      summary: Get vendor liabilities with aging
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vendor liabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiabilityAging"

  # ─── WALLETS ─────────────────────────────────────────────────
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List all wallets/accounts
      responses:
        "200":
          description: List of wallets
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Wallet"
                required: [wallets]
    post:
      operationId: createWallet
      tags: [wallets]
      summary: Create a new wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWalletInput"
      responses:
        "201":
          description: Created wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallets/{id}:
    get:
      operationId: getWallet
      tags: [wallets]
      summary: Get wallet details and statement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletDetail"
    patch:
      operationId: updateWallet
      tags: [wallets]
      summary: Update wallet balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                balance:
                  type: number
                name:
                  type: string
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"

  # ─── NOTIFICATIONS ───────────────────────────────────────────
...
components:
  schemas:
...
    Bill:
      type: object
      properties:
        id:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        description:
          type: string
        amount:
          type: number
        approvedAmount:
          type: number
        paidAmount:
          type: number
        outstandingBalance:
          type: number
        scheduledDate:
          type: string
        dueDate:
          type: string
        walletId:
          type: string
        walletName:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, approved, rejected, on_hold, partial, paid, overdue]
        createdBy:
          type: string
        createdByName:
          type: string
        hasAttachment:
          type: boolean
        overdueDays:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - vendorId
        - vendorName
        - description
        - amount
        - paidAmount
        - outstandingBalance
        - scheduledDate
        - priority
        - status
        - createdBy
        - createdByName
        - createdAt
...
    Vendor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        totalBilled:
          type: number
        totalPaid:
          type: number
        outstandingBalance:
          type: number
        lastPaymentDate:
          type: string
        createdAt:
          type: string
      required: [id, name, totalBilled, totalPaid, outstandingBalance, createdAt]
    VendorDetail:
      allOf:
        - $ref: "#/components/schemas/Vendor"
        - type: object
          properties:
            bills:
              type: array
              items:
                $ref: "#/components/schemas/Bill"
            recentActivity:
              type: array
              items:
                $ref: "#/components/schemas/AuditEntry"
          required: [bills, recentActivity]
    CreateVendorInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
      required: [name]
    LiabilityAging:
      type: object
      properties:
        vendorId:
          type: string
        vendorName:
          type: string
        totalOutstanding:
          type: number
        aging0to7:
          type: number
        aging8to14:
          type: number
        aging15to30:
          type: number
        aging30plus:
          type: number
      required: [vendorId, vendorName, totalOutstanding, aging0to7, aging8to14, aging15to30, aging30plus]
    Wallet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        balance:
          type: number
        currency:
          type: string
        isLow:
          type: boolean
        createdAt:
          type: string
      required: [id, name, balance, currency, createdAt]
    WalletDetail:
      allOf:
        - $ref: "#/components/schemas/Wallet"
        - type: object
          properties:
            recentTransactions:
              type: array
              items:
                $ref: "#/components/schemas/Bill"
          required: [recentTransactions]
    CreateWalletInput:
      type: object
      properties:
        name:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        balance:
          type: number
        currency:
          type: string
      required: [name, balance, currency]
...
    AuditEntry:
      type: object
      properties:
        id:
          type: string
        billId:
          type: string
        userId:
          type: string
        userName:
          type: string
        action:
          type: string
        details:
          type: string
        oldValue:
          type: string
        newValue:
          type: string
        createdAt:
          type: string
      required: [id, userId, userName, action, createdAt]
...